requirements.txt
appnope==0.1.0
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
freetype-py==2.1.0.post1
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.1
nose==1.3.7
numpy==1.16.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tqdm==4.35.0
traitlets==4.3.2
wcwidth==0.1.7
